requirements.txt
fastapi
uvicorn
requests
python-dotenv
langgraph
langchain
google-generativeai
